requirements.txt
streamlit>=1.30.0
openai>=1.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.18.0
Pillow>=10.0.0
plotly>=5.18.0
streamlit-extras>=0.3.6
google-generativeai>=0.3.0

# Voice Input Dependencies
audio-recorder-streamlit>=0.0.8
SpeechRecognition>=3.10.0
pydub>=0.25.1

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
